Gutted a large portion of the original code, to make maintenance and working with it easier for our purposes. This code is very well written, the original author deserves credit.
I've turned the previous Tantalum experience into a simplified interactive screensaver of sorts (disable the mouse and it would be just a screensaver).  Created a new scene with a variety of objects in interesting locations.

Removed parts about saving images etc.
Removed checks for browser incompatibility, we're not servicing the public at large and on top of that, most of the extensions being tested have become mainstream since this was originally written.